Texturas y cosas de EC

+++ src/Engine/Utils/Time.h
@@ -11,26 +11,26 @@ struct Time {
 private:
 	friend TimeManager;
 
-	static double _deltaTime;
-	static double _time;
+	static float _deltaTime;
+	static float _time;
 public:
 	/// @~english
 	/// @brief Time in seconds passed since last update
 	/// @~spanish
 	/// @brief Tiempo en segundos que ha pasado desde la anterior actualización
-	static constexpr double const& deltaTime = _deltaTime;
+	static constexpr float const& deltaTime = _deltaTime;
 
 	/// @~english
 	/// @brief Time in seconds passed since the start of the program
 	/// @~spanish
 	/// @brief Tiempo en segundos que ha pasado desde el comienzo del programa
-	static constexpr double const& time = _time;
+	static constexpr float const& time = _time;
 
 	/// @~english
 	/// @brief Time in seconds between every physic's operations' update
 	/// @~spanish
 	/// @brief Tiempo en segundos cada el que se evalúan las operaciones físicas
-	static constexpr double fixedDeltaTime = 0.02;
+	static constexpr float fixedDeltaTime = 0.02;
 };
 
 #endif
