+++ src/Engine/Utils/TimeManager.cpp
@@ -7,7 +7,7 @@ void TimeManager::init() {
 
 void TimeManager::update() {
 	std::chrono::system_clock::time_point current = std::chrono::system_clock::now();
-	Time::_deltaTime = std::chrono::duration<double>(current - _previous).count();;
+	Time::_deltaTime = std::chrono::duration<float>(current - _previous).count();;
 	Time::_time += Time::_deltaTime;
 	_previous = current;
 }
